Split module system tests
